# [Fleet] Delete installed package assets in chunks to avoid high memory consumption